More separation and cleanup.
